main.c: Do remote command line parsing
Creates subclass of GApplication to override local_comman_line().
By doing this we are able to pass all command line handling to the
remote instance. Before, only commands that were not handled locally
were passed via 'command-line' signal to the remote instance.
Due to this, for example no help output that shows remotely handled options
was available.

A later implementation might be more differentiated with local and
remote handling, and pass arguments selectively.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>