bundle.c: set size return parameter earlier in check_bundle()
If an error occured in check_bundle() the 'size' return parameter was
not yet set to an appropriate value, even when this was already
available. This leads to incorrect outputs of the squashfs size.

Assigning the value as soon as it is available will allow showing
correct results in case of other failures.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>