bootstate.c: do not set priority for marking slot good
Marking a slot good resulted in giving it a fixed priority of 20.
This may make the slot non-booting if it owned a higher priority before.

The behavior of setting slot priority to 0 for marking a slot bad
remains untouched by this change as it does not conflict with other
priorities.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>