test/config_file: create config file from string
Create the config file from built-in string instead of using the
test/system.conf file from the repository. This allows furhter tests to
be added without having to create dozens of config file variants for
testing in the repository.

The config files are created in a tmpdir and assigned to a fixture to be
handled by set_up and tear_down methods.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>