test/common: do not assert result of r_umount in test_umount
test_umount may be used in tear-down functions to assure that nothing
mounted is left over. For this it is a common case that an unmount will
fail because of already having properly unmounted the respective
directory.
This change returns the original return value of r_umount and thus
allows the calling functions to do the assert if required. This
additionally fixes the confusing behavior of always returning TRUE from
test_umount.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>