signature: obtain verbose error string for CMS_verify() failures
The genreic error message "signature error" is too unspecific to debug
issues with certificates. The Openssl library routines provide routines
to for getting more precise error codes and messages which are used
here.

A side-effect of this is that the openssl library does not print error
messages in its own format directly to screen, but only through the
normal rauc error reporting mechanism.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>