context: call context preparation explicitly
r_context_configure is now r_context_prepare. If context
is not set yet, r_context_prepare initializes it.
r_context() does not prepare the context anymore. An
explicit call of r_context_prepare is needed.

This prevents context creation in the middle of an
operation and thus no progress step conflicts evolve.

Signed-off-by: Bastian Stender <bst@pengutronix.de>