service: add support for the info method
As a counterpart to "rauc info" add support for the
D-Bus command "Info".
The method "Info" returns information from the manifest:
- compatible
- version

The usage here is that an application can query for the
compatible and version information and based on this
information allow a user to proceed or not with an update.

Signed-off-by: Sam Ravnborg <srn@skov.dk>