manifest: check for unparsed keys and groups
So far having an invalid key or group in the manifest was neither
discovered nor handled. This might make it more difficult for manifest
writers to discover issues. The only hint for a bug in their manifest was
that the respective key or group suddenly dissapeared in the rewritten
manifest.

This patch removes each parsed key of the currently parsed group from
the GKeyFile structure and reports an issue if there is a remaining key
after having parsed the entire group. Accordingly, each group is removed
after having parsed it. At the end, a remaining group will be reported,
too.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>