bundle: add sanity checks for signature size
If the input bundle contains broken data, signature size will be set to
an invalid size limited only by the data type (uint64).

As this size is used for memory allocation, it is very critical to use
this size parameter unchecked.

This patch introduces two sanity checks: the first validates that the
signature size does not exceed the size of the entire bundle, while the
second one checks if the signature size exceeds 64KiB which should be a
generous upper limit for signatures.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>