update_handler: use -F on mkfs.ext4 to allow working on files
If not given, some versions of mkfs.ext4 will complain that the slot
files used in test are "not a block special device". Providing the `-F`
option will prevent this warning.

Note that according to the man mkfs.ext4 page, givin the `-F` option
only once will not allow writing on mounted file systems.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>