install: slotstatuspath may not be set when trying to write status file
The variable slotstatuspath is initialized after mounting the target
slot when reading the slot status file for checking if we can skip the
update.
If mounting the target slot for reading the slot status path failed
(which is uncritical as it might have been a virgin or broken slot) but
the image update ran successfully and we attempt to write the new status
file, the variable slotstatuspath will not be set, yet.

Thus it has to be set after the second mount too in case it is still
NULL.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>