bundle: assert that no empty buffer will be passed to g_input_stream_read_all
The glib function g_input_stream_read_all() will return FALSE but not
set GError if its buffer argument is NULL. This will only happen is the
`count` argument passed to input_stream_read_bytes_all is 0.

Thus add this assert to make clear a value of 0 will lead to an error.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>