main: add --debug switch to turn on debugging output
This is a helper for debugging rauc and issues concerning the update
process. The result of adding `-d` or `--debug` to your rauc call is
similar to calling rauc with `G_MESSAGES_DEBUG=rauc` set.

Note that the debug switch handling takes into account any predefined
value of the G_MESSAGES_DEBUG environment variable.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>