install: control automatic activation of newly installed slot
Add boolean key "activate-installed" to section [system] of RAUC's "system.conf"
and use its value to decide if a slot shall be activated with respect to the
bootloader after being updated.

Extend the test suite with tests for correctly parsing the key's values
"true"/"false", for detecting typos in the value and for setting the correct
default value "true" if the key does not exist in the [system] section of the
config file at all.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>