test/config_file: add tests for boolean key "readonly"
Extend an existing test case to test for correctly parsing the value "false" and
add a new test for detecting typos in the value.

There have already been tests for correctly parsing the value "true" and for
setting the correct default value "false" if the key does not exist in the
slot's section of the config file at all.

Signed-off-by: Ulrich Ölmann <u.oelmann@pengutronix.de>