src/update_handler: allow uncompressd tarballs
Since update files are compressed in the bundle with squashfs, there is no need to
require them to be compressed prior to being bundles.

This commit removes the dot from the tar entries in the updatepairs
array, so that *.tar files are handled just like e.g. .tar.gz files.

Signed-off-by: Martin Hundebøll <mnhu@prevas.dk>