src/bundle: fix handling of rauc info --no-verify
--no-verify is meant to be a flag for disabling signature verification
check, not for entirely disabling bundle checking.

Despite the check_bundle() function already provides a 'verify' argument,
it was called only in case of 'verify' diabled. This is now replaced by
calling check_bundle() unconditionally.

For this, the check_bundle() function had to be fixed to honor the 'verify'
argument when deciding if a missing keyring file would break bundle
checking or not.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>