src/main: add keyring command line argument
When using rauc info, for example a keyring file is required in order to
be able to verify the bundle signature.
The keyring could only be configured from the system configuration file
which meant that you had to have a system.conf on your build host if you
wanted to use rauc info properly.

As a system.conf on the build host is not useful and could be
error-prone, thus a keyring argument for the RAUC CLI really simplifies
using rauc info and other possible commands requiring a keyring file.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>