src/install: support accessing remote bundles in check_bundle()
This patch enables parsing and handling of remote bundle locations, i.e.
prefixed by an URI scheme such as 'http://' or 'https://', as arguments
for the RAUC install and info.

At the moment, this is limited to bundle sizes of 8 MB as it is not
intended to be used for donwloading full-sized conventional bundles
containing disk images, yet.

The check_bundle() method will parse the URI scheme and download the
bundle to a temporary location that will be cleaned up when calling
free_bundle() on the returned RaucBundle structure.

The origin of the original bundle will be saved in newly introduced
`origpath` member of struct RaucBundle. This will also allow detecting
and properly handling temporary stored bundles in other parts of RAUC.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>