src: add is_slot_mountable() to not attempt to mount unmountable slots
Prior to this, checking if slot is writeable was done directly in the
launch_and_wait_default_handler() and only for writing slot status file,
not for reading.

Now a simple method allows to perform the check for all image types
supported and prevents from unnecessary warnings about mount/unmount
issues.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>