src/main: add resign command
When using RAUC in context of a full test and deployment infrastructure,
it might be a common case to let the build system infrastructure create
a bundle with a development key. This bundle will undergo certain tests
to verify it installs a working system, etc. Finally, after having
verified that, this bundle should be signed with a production key
while not touching its content anymore (might also play a role in the
context of certification).

This patch introduces the `resign` command which supports exactly
this. By calling `rauc resign` on an existing bundle, it will be
verified with the given keyring and resigned with the using the provided
key and certificate:

  rauc resign --keyring=... --key=... --cert=... in.raucb out.raucb

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>