src/bundle: always call check_bundle explicitly
Prior to this change, all functions that required to have a checked and
verified bundle called check_bundle on their own.

This had a couple of disadvantages, for example that parameters as for
enabling or disabling verification had to be passed explicitly to all
functions that call check_bundle(),

Now that we have the RaucBundle structure, this enables us to use
check_bundle() as a kind of open() function for the bundles, meaning
that you now have to call check_bundle() explicitly for gaining a
RaucBundle struct. All functions that formerly used check_bundle() now
only get the RaucBundle struct of the opened bundle passed.

This reduces the argument list for most functions while easing the
addition of further functionality and bundle parameters passed.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>