test/bundle: introduce bundle_fixture_set_up_bundle
Most tests rely on using a standard RAUC bundle and it on their own
based on the manifest data and content dir. The fixture could do this
task commonly for all test.

As there are also some tests that do not require any pre-generated
bundle data, we also split bundle fixtures up into one that generates a
dummy bundle and one that does not.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>