src/signature: add functions to generate hash of public key info
The hashes are meant to be able to securely compare certificates.

The default functions allow either to hash the entire certificates
(which is impractical as this also hashes e.g. the date info and thus
will not allow to compare over renewed certs) or only the public key
(which is imprecise and potentially more vulnerable as it omits some
important meta information as the algorithm used)

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>