src/install: split and rework target slot selection handling
This patch reworks the slot selection mechanism with these major goals:

* make target slot selection independent from the manifest
* simplify selection algorithm

The former approach had some disadvantages:
The function determine_target_install_group() which was responsible for
determining the appropriate target slots evaluated only those slots
that were contained in the manifest.
Despite the promise that a target slot mapping will be made available
also for the full custom handler, this will not work when having no
images specified in the manifest.
Additionally this handling is quite inflexible considering further
extensions of manifest image image specifications that will require some
selection of valid images.

This patch now changes the meaning of determine_target_install_group()
to only generate a target install group, i.e. a mapping of each slot
class to one individual member of this class that can be used to install
a slot.

Then the new method get_install_images() is introduced which walks
through the list of images specified in manifest, performs some checks
(such as if there is a target slot for each image) and generates a list
of images to install. By now this list should match the list of all
images also listed in the manifest, but it provides the option to
perform future filter operations on the actual target image list.

Installation handling is changed in order to iterate over the new
selected target image list instead of the original manifest image list.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>