src/config_file add support for system variant key
This adds they keys 'variant-dtb', 'variant-name' and 'variant-file' to
the [system] section in the RAUC system configuration file.

As its name suggests, the variant allows to specify a specific variant
of an updatable system handled by RAUC. This is a preparation for
introducing image variants that allow to place different 'variants' of an
image in the same bundle, e.g. for installing target-sepcific DTB or
kernel without having to build a separate bundle for each.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>