src/install: handle image variants during installation
This changes the get_install_images() method to honor image variants in
RaucManifest by selecting the appropriate variant for the system the
current installation runs on.

If there is an image for the 'default' variant, this will be valid for
each system, even if it has a specific variant configured.

BUT, if there is an image that exactly matches the sytems variant, this
one will be chosen instead of the default one.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>