src/signature: replace pointer-based assertion in cms_get_cert_chain()
Pointer equality of sk_X509_value(signers, 0) and
sk_X509_value(*verified_chain, 0) seems to be valid in some cases but
not, for example, for verified self-signed certificates (despite the
certificate content itself is equal).

This changes the assertion to be a comparison of both X509 structures
hashes which must be valid in each case.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>