docs/integration: More verbose and helpful description of u-boot support
Latest questions on both the RAUC GitHub issue tracker and the mailing
list have shown that the current U-Boot integration description is too
unverbose and lacks some important hints.

This patch should improve the situation by giving more hints on what is
required and what is expected for letting RAUC interact with U-Boot.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>