src/bootchooser: support multislot u-boot configurations
On a system using three slots (i.e. A, B, and R/recovery), the boot
order would be "A B R". When updating B from A, the bootorder is changed
to "B A", when it should be changed to "B A R".

Fix this by changing how the boot order environment string is
generated, so that the boot order is generated in steps:

  1. the current boot order is read with fw_printenv
  2. the new boot order is initialized with the updated slot
  3. the new boot order is appended with the current boot order, except
     the updated slot

Signed-off-by: Martin Hundebøll <mnhu@prevas.dk>