test/bootchooser: really test setting and getting barebox-state vars
This commit extends the dummy implementation of barebox-state used to
perform test by adding ability to test setting and getting of variables.

For this we can now specify the pre- and post- values the state should
have via environment variables and validate the RAUC barebox bootchooser
backend sets and reads them properly.

Also adapt custom_handler to be compatible with new barebox-state checking

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>