src/bootchooser: implement barebox_state_get
This reactivates the old inactive barebox_state_get_int() function and
reworks it in order to let it both obtain priority and remaining
attempts of a given slot in one call.
This will avoid timing overhead by unnecessarily often calling the
barebox-state program.

Note that we still leave the method commented out as it is not used yet
and we do not want to trigger compile errors.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>