src/bootchooser: barebox_set_primary() respect disabled slots
So far the barebox_set_primary() function did unconditionally set the
primary slot to prio 20 and all other to prio 10.
This simple approach does not respect a disabled slot, that was set to
prio 0, e.g. by a 'bad' status mark.

To allow handling this, we need not only to write the barebox state but
also read it before. This allows us to detect inactive slots and
properly skip their reactivation.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>