src/main: output boot information in rauc status
This adds information about the bootloader backend to the output of
'rauc status'. It both lists which is the currently activated boot slot
(as a result of an update or manual execution of 'rauc status
mark-active') and the boot status of each bootable slot as either good
or bad (i.e. if it is marked bad, has remaining attempts, is listed in
priority list, etc.)

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>