src/update_handler: add support for writing ubifs to ubifs
Up to now, we only supported ubifs images for ubivol slot type.
Handling ubifs to ubifs slot type is quite similar, with the difference
that we assume the slot to be mountable when executing pre- or
post-install hooks and thus use mount_and_run_slot_hook() instead of
run_slot_hook().

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>