src/manifest: remove redundant check_compatible() function
Having this check in verify_manifest is not only redundant to what
verify_manifest() in install.c does, it is also a bug with respect to
the 'install-check' hook: Also if the hook will allow inequal
compatibles (which can be a proper case depending on the respective
requriements) the additional check performed by verify_manifest() will
reject the bundle.

The obvious solution for that is to simply remove the check_compatible()
call.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>