src/manifest: remove unused 'signature' argument from verify_manifest()
The verify_manifest() function had a 'signature' argument that aimed to
verify manifest itself, probably formerly intended to be used in
interaction with old network support.

There is no current method call of verify_manifest() in RAUC that uses
this argument (except the tests) in terms of having 'signature' set to
TRUE. Thus it is safe to remove this piece of dead code.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>