src: verify image checksums only right before image installation
The checksums contained in the manifest are slightly redundant with
respect to the signed bundle content. Thus, we do not really require
them for verification of the bundle content. But, despite that a
comparison between the images checksum and the pre-calculated one is
quite useful for sanity checking.

But, to not limit custom handlers and as preparation for future
approaches, we should not verify each image before being sure we want to
install them. Thus, this patch moves verify_checksum() call to the
launch_and_wait_default_handler().

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>