src/update_handler: add support for casync blob index extraction
This extends write_image_to_dev() with special handling for casync
.caibx files.

This allows addding the img_to_raw_handler and img_to_fs_handler for
slot types that allow directly writing to their device nodes and do not
require special handling such as ubifs (ioctl) or nand (nandwrite).

We use the active slot classes member as the seed for extracting the
image. Normally, in update scenarios the currently running system should
be closest to the target system and thus fit well for seeding.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>