src/bootchooser: support re-sorted EFI boot entries in efi_get_primary()
Some EFI implementations might choose to have a default EFI boot entry
always placed on first position in order to execute it first and then
continue selecting the next boot target to boot the actual system.

In this case, the simple approach on looking at the first element in
EFI BootOrder for finding the primary slot will not work.
Also having BootNext set must not necessarily mean that this will match
one of the bootnames handeled by RAUC.

A more generic approach for this would be to scan the entire BootOrder
list until finding the first element we know as a bootname for any of
our slots configured and consider this to be the primary one.

The patch changes the efi_get_primary() method to gain this behavior.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>