test/install-fixture: add always mounted bootloader slot
In order to test the "allow-mounted" feature for updating slots
mounted by the system at boot, such a "bootloader" slot is added to the
default test configuration.

Since an additional slot is added, the test for correct number of slots
in test/service.c is updated from 5 slots to 6 slots.

Signed-off-by: Martin Hundebøll <mnhu@prevas.dk>
Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>