src/main: do not access context from status printout functions
This introduces the local struct RaucStatusPrint as a container for
providing the information that the different formatter require to
assemble the output text in their respective format.

This decouples the formatta from the actual source of the data and thus
serves as a preparation for obtaining status data fully over D-Bus.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>