src: preserve original uig/gid during tar extraction
Always call tar extract with --numeric-owner option to preserve original
user and group IDs as packed into the archive.
This is required to assure reproducible and deterministic IDs that are
also consistent with the password file and prevent from later user
relabeling that would cause issues in the context of trusted boot.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>