Introduce max-bundle-download-size config setting (fixes #287)
At the moment, a hard-coded BUNDLE_DL_MAX_SIZE define is
used for a safety check when a bundle is downloaded.
To be more flexible at runtime, introduce a new config
parameter 'max-bundle-download-size'.
The current define is renamed to reflect the new meaning
better and used as default value when the config setting
is not specified.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>