Implement uboot_get_state() and uboot_get_primary()
This implements the r_boot_get_state() and r_boot_get_primary()
interface for the U-Boot bootloader.

The expected handling is implemented similar to what other bootloader
do. Both the BOOT_ORDER and BOOT_x_LEFT variables are taken into account
to determine the state of a slot and the primary one.

Signed-off-by: Timothy Lee <timothy.ty.lee@gmail.com>
Signed-off-by: Enrico Joerns <ejo@pengutronix.de>