src/signature: allow to use signature signing time for verification
In some security models it might be desired to use a different validity
model for signature verification.

While the default one used by OpenSSL and RAUC is to check if the
signature is valid at the time of verification, an alternative approach
is to check if the certificates were valid at the time of signing.

The newly introduced option 'use-bundle-signing-time' allows to select the
latter behavior if required.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>