src/bootchooser: uboot: remove slot from BOOT_ORDER for state bad
To be consistens with other implementations and to really disable
booting a slot on setting state bad, we now also remove the disabled
slots form the BOOT_ODER variable.

The behavior of uboot_get_state() already reflects this.

The tests need to be slightly adapted to reflect the new behavior.

Signed-off-by: Enrico Jorns <ejo@pengutronix.de>