src/main: get status information fully over D-Bus in service mode
By default, i.e. when compiled with --enable-service, we have a
separation of RAUC into the service controlling the installation and the
command line interface used to control it.
The communication between these two works via D-Bus as it would be for
any other client or frontend.

While for `rauc install` there is a clear seperation according to the
handling described above, for `rauc status` this was always not that
strict but more or less a 'hybrid' solution, mainly due to histroical
reasons. The (detailed) status information the client retrived via D-Bus
while all other information was examined by calls from the client, not
the server.
This not only breaks the design pattern but also adds some drawbacks
such as that the client also had to read the RAUC configuration file in
ordert to obtain the slot configuration and it did unnecessarily make its on
interaction with the bootloader, etc. which is a potential conflict with
what the service does at the same time.

After some pre-work this patch now finally fully implements the client
mode for rauc status and obtains all information via D-Bus only.
Only for the non-service case, i.e. without D-Bus the old behavior will
be retained.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>