src/install: actually handle readonly slots
So far we had the slot config booloan key `readonly` that allows to mark
a slot as being detected but not being writable by RAUC.

But, this option was only of informational nature and did not really
prevent from accidently installing content to a slot marked as readonly.

This patch now slightly extends get_install_images() to test for the
readonly argument if an image for the respective slot is to be installed
and aborts with an explicite error in this case.

This way it will not be possible to install data to 'readonly' slots,
but they are still considered e.g. for parent slot detection, etc.
An extra test to assure this was added.

Fixes #161

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>