include/signature: Explicitly document print_cert_chain() output parameters
This adds the doxygen tags [out] to those parameters that are made for
returning something.

Additionally, we add the tag [transfer full] for marking that the
returned object is newly allocated and must be freed afterwards.
This simply mimics glib convention of return type documentation.
Note that, as in glib, we omit transfer type in well-known parameters
such as GError return locations.

This is an exemplary change that could be applied to all other
functions, too.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>