src/signature: use newer get0 function names for obtaining times
In newer OpenSSL versions the function names
X509_get_notBefore/X509_get_notAfter are deprecated and replaced by
corresponding get0 functions to be clear about memory ownership.

To gain the benefit of this renaming be use the new names and define
the old ones for backwards compatibility.

This also requires to add some const to exisiting functions.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>