src/update_handler.c: fix 'preserve existing error' logic
mount_and_run_slot_hook is only called from within update_handler.c, and
always with the address of a stack variable as error.

As for archive_to_{ubifs,ext4,vfat}_handler, they are only called via a
function pointer returned from get_update_handler(). Checking the
callers of that in src/main.c and src/install.c show that they similarly
both pass &ierror as the error parameter.

Hence, in all these cases, "if (error)" is a tautology, which is
unlikely to what was intended. Per feedback from ejoerns, check upfront
that error is either NULL (meaning, the caller doesn't care about the
error) or that *error is initialized to NULL. Then rewrite the condition
to if (error && *error), which can now reliably check whether any of the
preceding logic had already set *error to some non-NULL value.

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>