src/bootchooser: fix segfault in efi_get_primary()
The variable bootnext might be null, so we shoud check it before trying
to access it. Plus, the error message seems to be inacurrate, it looks
like a copy/paste mistake or something.

This commit rewords the error message to make it correct, and fixes the
segfault at the same time.

Signed-off-by: Arnaud Rebillout <arnaud.rebillout@collabora.com>