src/install: inverse logic for network mode handling
Up to now, we assumed that a .raucb extension will mean having a bundle
to install and handled all other (or none) extensions as manifests for
network mode.

This would have lead to

  rauc install file:///path/to/bundle.rauc

being interpreted as network mode.

Since network mode is very unconvenient in RAUC and only partly
maintained, it is really not useful to assume it by default here.

As a result of this, we now explicitly make the `.raucm` manifest file
name extensions the special case and handle all other files as normal
RAUC bundles.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>