src/install: add escape hatch for already mounted devices
It may be useful to use the RAUC update mechanism to update slots representing
devices that are usually mounted. For example, on the RPi and Beaglebone Black
boards where both U-boot and its environment is stored in a VFAT filesystem,
that partition is necessarily mounted at boot.

There's currently no generic way to handle such a case, so this patch allows
one to mark a slot with "allow-mounted=true", with the caveat that it can only
be updated via a custom install hook.

While updating the bootloader in the above scenario is necessarily a dangerous
operation, this scheme can also be used for doing certain atomic and safe
updates to a live system, provided the mounted file system provides sufficient
atomicity guarantees (e.g. an atomic rename).

Signed-off-by: Rasmus Villemoes <rasmus.villemoes@prevas.dk>
Signed-off-by: Jan Luebbe <jlu@pengutronix.de>