src/main: a progress bar for RAUC
A little gadget I always wanted to have... And did in the train locally
because of poor german LTE performance...

It gives you a two-lined progress bar while installing with RAUC. The
first line contains the message in the same style as before, the second
line contains a hash-printing progress bar in the width of the current
terminal.

Output will be similar to this:

|  50% Checking Bundle done.
| [xxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxxx                                   ] 50%

This is only activated if we detect a tty, otherwise we use the
old-style printing format.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>