include/utils.h: add function to log subprocess calls
In various debugging scenarios it is quite useful to know which
subprocess calls RAUC actually initiates.

Often, one only get an exit code from the program, and maybe some more
or less helping error output. But for tracing down the issue and for
reproducing it, maybe also manually, it would be worth knowing how the
subprocess was actually called, i.e. with which pre-commands (sudo,
fakeroot, etc.) and with which arguments.

This adds a new debug domain 'rauc-subprocess' and a corresponding
helper function for dumping subprocess arguemnt calls that are often
prepared in the form of a GPtrArray.

To get output of this, run rauc with

  G_MESSAGES_DEBUG=rauc-subprocess rauc ...

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>