src/main: fix G_MESSAGES_DEBUG handling for debug switch
The former check `!g_strrstr(domains, G_LOG_DOMAIN)` matches on every
occurrence of G_LOG_DOMAIN ('rauc') in G_MESSAGES_DEBUG which also leads
to not appending 'rauc' log domain when 'rauc-subprocess' log domain is
already defined. This is not intended.

Removing this check completely will potential result in adding
G_LOG_DOMAIN twice if it was passed by G_MESSAGES_DEBUG environment
variable already, but that is not a remarkable drawback.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>