src/main: more targeted log domain debugging output
Only printing "Domains: " for outputting the used log domains might be
misleading. Thus we print "Debug log domains" instead.

It should also be emitted via logging functions, not via g_print as it
is no real user output. Note that we should not use g_debug here, as
this will not show up if the default log domain is not included.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>