src/context: introduce r_context_begin_step_formatted()
This allows printing progress information with printf-like format
strings.
This is a convenience function to not require allocating and freeing
strings for each step call.

Note that by using this function in launch_and_wait_default_handler()
where g_strdup_printf() was used before to construct the string we also
fix a memory leakage.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>