src/signature: use OPENSSL_config() instead of OPENSSL_no_config()
This will now load OpenSSL configuration from openssl.cnf default
section 'openssl_conf'.

This might be useful, anyway, but the main reason we have to change this
is to stay consistent with upcoming support for OpenSSL 1.1.1.

We also link again libcurl and libcurl uses OPENSSL_init_ssl() which
itself calls OPENSSL_init_crypto() with OPENSSL_INIT_LOAD_CONFIG.
Now, due to a bug/feature in OpenSSL, it is no longer possible to initialize
OpenSSL with both 'load config' and 'no config' (refer
https://github.com/openssl/openssl/issues/7350 for details and
discussion).
This forces us to initialize OpenSSL with 'load config', too, when we want
to support OpenSSL 1.1.1.

This is a preparation that changes the old code to gain the same
behavior that future OpenSSL-1.1.1-supporting code will have as we do
not want to behave differently with different OpenSSL versions.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>