src/service: allow to return failure form r_service_run()
So far, even if a call of 'rauc service' failed with

| Failed to obtain name de.pengutronix.rauc

the service returned exit code 0, indicating everything is fine.

The main reason why this was not handled properly is that
r_service_run() unconditionally returned 0 after its main loop got
terminated. As we cannot simply return a value when terminating the main
loop with 'g_main_loop_quit()', this introduces the variable
'service_return' that we pass to the bus handlers and that
r_on_name_lost() will set to FALSE in case of an error.

By returning this variable instead of fixed TRUE, the service now aborts
correctly with indicating a failure if, e.g., the name could not be
obtained on the bus.

Signed-off-by: Enrico Joerns <ejo@pengutronix.de>